Rework string parser to use a single table.
This is because it turns out section names aren't necessarily ordered in the
string table, so we need to look them up by the actual byte index (eww). So,
we get rid of the section name table, the fine name pointer, and add an array
of ints: these correspond to a member in the string table, and identify what
index they had in the in-file table.