Rework ingest_link_order, us canonical symtab and relocs
Using the internal coff reloc struct was pretty painful beforehand, and when
we start dealing with symbol table indexes, downright impossible. So instead
use the canonical version, which the backend should happily extract for us.
This also solves the problem of not knowing whether the input section is coff
or doff (see previous erased comments about this situation).