Set out struct/function definitions for how I want to reorganise operand beating
Right now it's all horrificly unpleasent. Responsibility for different
bits of operation are scattered all over the place, and theres a lot
of code replication (the optest stuff) to try and cope with guessing
what instruction template to use. The goal of working out the unit and
side without the need for the specifier is pretty impossible.

So instead this aims to split different portions of beating an operand
into different utility functions. There's actually parsing/reading
the line of text into some operand storing record; validating that an
operand is valid for a particular unit/side specifier; and emitting
the operand data to a binary opcode.

Also, bees.