Remunge operand handler strategy: single handler->multi types, not multi->multi.
This is because it was going to be unpleasent working out what type
operand we have, when there were several that could match: for
example, which out of srcreg1, srcreg2, or dstreg should it be?

The answer is, we shouldn't care at the time when we're parsing the
line, so instead we'll just have a list of the different text types
it _might_ be, but which are pretty much indistinguishable. This
greatly reduces some of the logic required to start picking out
instructions.