Rework and simplify how we represent unitspec validity.
The previous method was to store validity flags for each unit and for
each side; however on second thought it shouldn't make a difference
which unit the instruction executes on, and if it did, then it should
be a different instruction template.

So that goes out the window; We retain flags showing whether the the
template will work for sides 1 or 2, whether it'll need the xpath to
do that. We also keep the "uses data path 2" flag, but it's still
slightly out of place, as it should be a "overall instruction" kind
of flag, rather than template specific.