Remove disgusting old unit-finding code, replace with new version.
This means the horrific juggling and macro situation used before all
disappears, and we instead place a call to the recursive routines
that attempt to allocate units to all instructions of different
forms.

With this commit we lose the code that invokes calls to assemble mv
instructions, this is something else that will have to be rewritten
too.