Spit out errors/warnings if constant offset is not a sane kind of expression.
The feature here is that if gas couldn't parse it, or it didn't
exist then we'll complain. If however you use a symbol based expr
as an offset to memory access, you're A) UTTERLY NUTS and B)
likely to hit an untested code path.

In reality it isn't nuts to do that - TIs ABI specifies a register
(is it B14?) as the "data page" register. Presumably this points at
the base of .data, and you can happily access the data area without
having to load an absolute base pointer first; the offset into that
has to be relocated in case you link multiple .data sections together
and offsets into it change.

This is something to pump into the compiler at some point.