Remove scale_output memaccess operand field.
On second thoughts, we can't communicate this to the operand
writer via the read_operand struct, as other validators might
run in the meantime and overwrite this field to an invalid value.
Instead, the output writer should be able to determine that it
needs to scale the offset by itself.