Rework how invalid operands are picked out for error printing.
Essentially the problem here is that the operand at fault was being
recorded, but the side that faulted it was not. So, introduce a new
set of flags, one for each operand on each side. Naturally this means
more fluff, but we can now pick out the correct faulting operand.