Once a unit is picked for an insn, mark insn as _only_ executing on that unit.
The problem here is best illustrated by nop, which can execute on
anything - if it's the first (or only) instruction being unit selected,
there is nothing to cause its other unit options to be masked out.
Which means no single unit gets selected. Which leads to trouble.

Thanks to valgrind for spotting a number of problems orientated
around the aforementioned trouble.