Fix relocations when linking coff2 files:
Essentially BFD is a twisty turn maze of adding constants to the addend
"To compensate for when we do it again later", all alike. Instead of
trying to honour all of this, assume that the addend is 0 all of the
time, and provide our own reloc installer that takes into account any
addend already encoded in the insn.

Without this, whenever coff files get linked together, we end up with
the relocations being against a symbol minus the offset from that
symbol, because of all this addend juggling (it should just be zero).
It appears that bfd wasn't reading the addend out of the instruction
data at all.