Disable array boundry checking:
Unfortunately include/coff/ti.h's PUT_SCNHDR_FLAGS macro happens to
violate a bunch of array boundries, in lieu of using offsets. It's
too in-depth to actually perform a fix, and the code is still
present in CVS, so binutils probably haven't fixed it either. As a
workaround, we just disable array boundry checking.

Hacks, on top of rafts of hacks.